finalized my thoughts.
I took the chakra calendar off the page. I may attach a link to a separate page with the calendar on that. I added the prompt feature to my design. I have to fix a few things, such as the alert pops up when I use the navigation bar. I also need to fill my nav bar in with real and relevant information.