updated the javascript
I noticed that if I didn't add a name, nothing happened, which was confusing. So, I created a prompt, and if the user doesn't enter a name, they are alerted, and then the prompt comes back until they enter a name. The name is necessary to make the rest of the page work as intended.